refactor: correct referenicing group and post modals from pages
