+++ client/src/pages/groups/index.tsx
@@ -6,7 +6,7 @@ import PlatformLayout from '~/layouts/platform';
 import useStore from '~/stores/utils/useStore';
 import { useUserStore } from '~/providers/store-providers/userStoreProvider';
 import { type UserState } from '~/stores/userStore';
-import AddGroupModal from '~/app/_components/groups/add-group-modal';
+import AddGroupModal from '~/app/_components/groups/AddGroupModal';
 
 export default function Groups() {
 	const groupId = '69';
